refactor(core): big refactor
