+++ lua/adev/lsp/servers/emmet_ls.lua
@@ -1,0 +1,21 @@
+return function()
+    vim.lsp.config("emmet_ls", {
+        filetypes = {
+            "astro",
+            "css",
+            "eruby",
+            "html",
+            "htmldjango",
+            "javascriptreact",
+            "less",
+            "pug",
+            "sass",
+            "scss",
+            "svelte",
+            "typescriptreact",
+            "vue",
+            "htmlangular",
+            "blade",
+        },
+    })
+end
