+++ lua/adev/lsp/servers/html.lua
@@ -1,0 +1,12 @@
+return function()
+    vim.lsp.config("html", {
+        filetypes = {
+            "astro",
+            "eruby",
+            "html",
+            "htmldjango",
+            "templ",
+            "blade",
+        },
+    })
+end
