+++ lua/adev/lsp/servers/intelephense.lua
@@ -1,0 +1,10 @@
+return function()
+    -- NOTE: a secure way to handle licence key
+    vim.lsp.config("intelephense", {
+        init_options = {
+            licenceKey = vim.fn
+                .system({ "cat", os.getenv "HOME" .. "/.secrets/intelephense" })
+                :gsub("\n", ""),
+        },
+    })
+end
