+++ lua/adev/lsp/servers/pylsp.lua
@@ -1,0 +1,11 @@
+return function()
+    vim.lsp.config("pylsp", {
+        settings = {
+            pylsp = {
+                plugins = {
+                    pycodestyle = { ignore = { "W391" }, maxLineLength = 120 },
+                },
+            },
+        },
+    })
+end
