+++ lua/adev/lsp/servers/tailwindcss.lua
@@ -1,0 +1,16 @@
+return function()
+    vim.lsp.config("tailwindcss", {
+        settings = {
+            tailwindCSS = {
+                experimental = {
+                    classRegex = {
+                        -- cva("…")
+                        { "cva\\(([^)]*)\\)", "[\"'`]([^\"'`]*)[\"'`]" },
+                        -- cn("…")
+                        { "cn\\(([^)]*)\\)", "[\"'`]([^\"'`]*)[\"'`]" },
+                    },
+                },
+            },
+        },
+    })
+end
