+++ lua/adev/lsp/servers/vue_ls.lua
@@ -1,0 +1,10 @@
+return function()
+    vim.lsp.config("vue_ls", {
+        init_options = {
+            typescript = {
+                tsdk = os.getenv "HOME"
+                    .. "/.local/share/nvim/mason/packages/vue-language-server/node_modules/typescript/lib",
+            },
+        },
+    })
+end
